Updated disclaimerPage
changed it and disclaimer_lang(now disclaimerLanguageSelect) to have better readability
