+++ lib/disclaimerPage.dart
@@ -7,7 +7,7 @@ import 'package:mazilon/util/styles.dart';
 import 'package:mazilon/util/appInformation.dart';
 import 'package:mazilon/util/userInformation.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:mazilon/util/disclaimer_lang.dart';
+import 'package:mazilon/util/disclaimerLanguageSelect.dart';
 
 // the disclaimer page widget,
 // it shows the disclaimer text and a button to confirm the disclaimer
@@ -57,25 +57,7 @@ class _DisclaimerPageState extends State<DisclaimerPage> {
             child: Center(
               child: Column(
                 children: [
-                  SizedBox(height: 20.0),
-                  Container(
-                    width: MediaQuery.of(context).size.width > 1000
-                        ? 600
-                        : MediaQuery.of(context).size.width * 0.6,
-                  ),
-
                   LanguageDropDown(changeLocale: widget.changeLocale),
-                  SizedBox(height: 20.0),
-                  Container(
-                    width: MediaQuery.of(context).size.width > 1000
-                        ? 600
-                        : MediaQuery.of(context).size.width * 0.6,
-                  ),
-
-                  LanguageDropDown(changeLocale: widget.changeLocale),
-                  SizedBox(
-                      height:
-                          20.0), //space between the top of the screen and the disclaimer text
                   Padding(
                     padding: const EdgeInsets.fromLTRB(20, 0, 20, 10),
                     child: myAutoSizedText(
